requirements.txt
discord.py
flask
python-dotenv
youtube-dl